Update regex to catch FIQhack
Some variants report a character name alongside the username of the
person playing; this commit updates the Hardfought/Beholder regex to
match these message types, in such a way that they will work with
always_allow_users. Previously they didn't match the regex and so got
shown in the main buffer regardless of what rules the user had applied.